refact: create a separated file for managing the custom application rule preferences
